cdk/iam-manager/requirements.txt
-e .
aws-cdk.aws-lambda
aws-cdk.aws-apigateway
aws-cdk.aws-s3
aws-cdk.aws-glue
aws-cdk.aws-cloudtrail
aws-cdk.aws-athena

